Restore mod priority configuration option for Orechid outputs
If the configured list is non-empty, the first namespace ID on the list that is matched to blocks in the potential output tag is used to determine the output. If more than one such block is found, the output is randomized amon those blocks. If none of the namespace IDs matches any output block, the result is picked randomly as if no priority mods were configured.